chore(clients/go): moved zbc into pkg
